Move navigation components to their own files
