+++ src/App.tsx
@@ -3,7 +3,7 @@ import { KeyboardAvoidingView, Platform } from "react-native";
 import { Provider } from "react-redux";
 import store from "./state/store";
 import { SharedRefetchProvider } from "./sharedRefetchContext";
-import Navigation from "./navigation/AppNavigator";
+import AppNavigator from "./navigation/AppNavigator";
 
 LogBox.ignoreLogs(["The useProxy option is deprecated"]);
 
@@ -17,7 +17,7 @@ export default function App() {
           style={{ flex: 1 }}
           behavior={Platform.OS === "ios" ? "padding" : undefined}
         >
-          <Navigation />
+          <AppNavigator />
         </KeyboardAvoidingView>
       </SharedRefetchProvider>
     </Provider>
