+++ src/screens/EditProfile.tsx
@@ -14,7 +14,7 @@ import { RootState } from "@/state/store";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import handleError from "@/utils/handleError";
 
-export default function RegisterScreen() {
+export default function EditProfile() {
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
   const [email, setEmail] = useState("");
